requirements.txt
httpx>=0.27.0
websockets>=12.0
pandas>=2.2.0
numpy>=1.26.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiosqlite>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
